refactor: replace old PeopleSelectorModal everywhere and rename ProposalPeopleSelectorModal to PeopleSelectorModal
- Convert instrument and technique scientist assignment to the new
  autocomplete-based selector (no email invites)
- Remove the old table-based PeopleSelectorModal
- Rename ProposalPeopleSelectorModal to PeopleSelectorModal since it is
  no longer proposal-specific
- Update e2e tests for the new selector flow (instruments, techniques,
  FAP member assignment in event log test)
